relion_refine: added --keep_scratch

+++ src/ml_optimiser.cpp
@@ -397,6 +397,7 @@ void MlOptimiser::parseContinue(int argc, char **argv)
 	fn_scratch = parser.getOption("--scratch_dir", "If provided, particle stacks will be copied to this local scratch disk prior to refinement.", "");
 	keep_free_scratch_Gb = textToInteger(parser.getOption("--keep_free_scratch", "Space available for copying particle stacks (in Gb)", "10"));
 	do_reuse_scratch = parser.checkOption("--reuse_scratch", "Re-use data on scratchdir, instead of wiping it and re-copying all data. This works only when ALL particles have already been cached.");
+	keep_scratch = parser.checkOption("--keep_scratch", "Don't remove scratch after convergence. Following jobs that use EXACTLY the same particles should use --reuse_scratch.");
 
 #ifdef ALTCPU
 	do_cpu = parser.checkOption("--cpu", "Use intel vectorisation implementation for CPU");
@@ -639,6 +640,7 @@ void MlOptimiser::parseInitial(int argc, char **argv)
 	fn_scratch = parser.getOption("--scratch_dir", "If provided, particle stacks will be copied to this local scratch disk prior to refinement.", "");
 	keep_free_scratch_Gb = textToInteger(parser.getOption("--keep_free_scratch", "Space available for copying particle stacks (in Gb)", "10"));
 	do_reuse_scratch = parser.checkOption("--reuse_scratch", "Re-use data on scratchdir, instead of wiping it and re-copying all data.");
+	keep_scratch = parser.checkOption("--keep_scratch", "Don't remove scratch after convergence. Following jobs that use EXACTLY the same particles should use --reuse_scratch.");
 	do_fast_subsets = parser.checkOption("--fast_subsets", "Use faster optimisation by using subsets of the data in the first 15 iterations");
 #ifdef ALTCPU
 	do_cpu = parser.checkOption("--cpu", "Use intel vectorisation implementation for CPU");
@@ -2362,12 +2364,13 @@ void MlOptimiser::iterateWrapUp()
 {
 
 	// delete barrier, threads and task distributors
-    delete global_barrier;
+	delete global_barrier;
 	delete global_ThreadManager;
-    delete exp_ipart_ThreadTaskDistributor;
+	delete exp_ipart_ThreadTaskDistributor;
 
-    // Delete volatile space on scratch
-    mydata.deleteDataOnScratch();
+	// Delete volatile space on scratch
+	if (!keep_scratch)
+		mydata.deleteDataOnScratch();
 
 #ifdef MKLFFT
 	fftw_cleanup_threads();
